WIP(screener): local extensions (universe, filters, signals, fetch, analyze_history, config, README, etc.) + .gitignore runtime ignores
Preserves local work (stabilize + further dev on top of 37560ce) before merging remote changes from origin/main (tracking.py, track_performance.py, backfill, multi-universe polish, forward win-rate, full ranked output etc. from parallel Kimi session).

+++ hydra_screener_local/core/regime.py
@@ -32,14 +32,12 @@ def compute_rich_regime_scores(
     if len(spy) < lookback:
         return RegimeScores(0.5, 0.5, 0.5, 0.5, 0.5, 0.5)
 
-    current = spy.iloc[-1]
-
-    # 1. Trend (SPY vs SMA200 + pendiente reciente)
-    sma200 = spy.rolling(lookback).mean().iloc[-1]
+    current = float(spy.iloc[-1])
+    sma200 = float(spy.rolling(lookback).mean().iloc[-1])
     trend_strength = 1.0 if current > sma200 else 0.0
     
     # Pendiente de 20 días normalizada
-    ret_20 = (current / spy.iloc[-20] - 1) if len(spy) >= 20 else 0
+    ret_20 = (current / float(spy.iloc[-20]) - 1) if len(spy) >= 20 else 0
     trend_strength = (trend_strength * 0.6) + (np.clip(ret_20 + 0.04, -0.1, 0.15) / 0.15 * 0.4)
 
     # 2. Volatility Regime (vol actual vs vol histórica)
@@ -49,13 +47,13 @@ def compute_rich_regime_scores(
     volatility_score = np.clip(1 - (vol_ratio - 0.8) / 1.2, 0, 1)  # menor = mejor (menos vol)
 
     # 3. Momentum Strength (combinación de 20d y 60d)
-    ret_60 = (current / spy.iloc[-60] - 1) if len(spy) >= 60 else 0
+    ret_60 = (current / float(spy.iloc[-60]) - 1) if len(spy) >= 60 else 0
     momentum_score = np.clip((ret_20 * 0.6 + ret_60 * 0.4 + 0.05) / 0.20, 0, 1)
 
     # 4. Drawdown Velocity (qué tan rápido fue el drawdown reciente)
     rolling_max = spy.rolling(60).max().iloc[-1]
     current_dd = max(0.0, (rolling_max - current) / rolling_max)
-    dd_20d_ago = max(0.0, (rolling_max - spy.iloc[-20]) / rolling_max) if len(spy) >= 20 else current_dd
+    dd_20d_ago = max(0.0, (rolling_max - float(spy.iloc[-20])) / rolling_max) if len(spy) >= 20 else current_dd
     velocity = max(0.0, current_dd - dd_20d_ago) / 20.0   # velocidad promedio por día
     velocity_score = 1.0 - np.clip(velocity * 40, 0, 1)    # alta velocidad = peor
